+++ threedi_models_and_simulations/widgets/new_schematisation_wizard.py
@@ -21,6 +21,7 @@ from ..utils import EMPTY_DB_PATH, SchematisationRasterReferences, extract_error
 from ..utils_qgis import geopackage_layer
 from ..utils_ui import ensure_valid_schema, get_filepath, read_3di_settings, save_3di_settings, scan_widgets_parameters
 
+
 base_dir = os.path.dirname(os.path.dirname(__file__))
 uicls_schema_name_page, basecls_schema_name_page = uic.loadUiType(
     os.path.join(base_dir, "ui", "new_schematisation_wizard", "page_schema_name.ui")
@@ -65,7 +66,10 @@ class SchematisationNameWidget(uicls_schema_name_page, basecls_schema_name_page)
         """Return new schematisation name, tags and owner."""
         name = self.le_schematisation_name.text()
         description = self.le_description.text()
-        tags = self.le_tags.text()
+        if not self.le_tags.text():
+            tags = []
+        else:
+            tags = [tag.strip() for tag in self.le_tags.text().split(",")]
         organisation = self.cbo_organisations.currentData()
         owner = organisation.unique_id
         return name, description, tags, owner